adk_service/requirements.txt
google-adk>=0.5.0
litellm>=1.50.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
pydantic>=2.9.0
aiosqlite>=0.20.0
truststore>=0.9.0
